rename [mudar o nome] s7.c .h to aboa-s7.c .h prior to changing any code [antes de alterar qualquer código], to ease diffing [para facilitar a comparação]
